refactor: extract shared MCP tool name formatting into utils/toolName.ts
Address AI review: consolidate duplicate shortToolName (ConversationView)
and displayToolName/parseMcpName (ToolCallBlock) into a single shared
utility. Add clarifying comment on scanner's separate MCP tracking path.
